v1.1: significantly improve project structure, still runs all well and fine
