+++ java/src/console/utils/ValidCommands.java
@@ -1,13 +1,14 @@
-package Settings;
+package console.utils;
 
-
-import java.util.*;
-import java.util.stream.Collectors;
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.LinkedHashSet;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 public class ValidCommands {
-
   // --- Command definitions ---
   private static final Set<String> menuCommands = Set.of(
           "start", "delete"
@@ -53,4 +54,3 @@ public class ValidCommands {
     return Collections.unmodifiableSet(all);
   }
 }
-
