--- src/console/Settings/GameSettings.java
@@ -1,4 +1,0 @@
-package Settings;
-
-public class GameSettings {
-}
